+++ api/models/clip.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, ForeignKey
+from sqlalchemy import Column, Integer, Float, String, ForeignKey
 from sqlalchemy.orm import relationship
 
 from .base import Base, TimestampColumns
@@ -10,7 +10,7 @@ class Clip(Base, TimestampColumns):
     id = Column(Integer, primary_key=True)
     name = Column(String, nullable=False)
     video_id = Column(Integer, ForeignKey("videos.id"), nullable=False)
-    start_at = Column(Integer, nullable=False)
-    end_at = Column(Integer, nullable=False)
+    start_at = Column(Float, nullable=False)
+    end_at = Column(Float, nullable=False)
 
     video = relationship("Video", back_populates="clips", lazy="joined")
